refactor(frontend): [FRO-009] :sparkles: Move all interfaces to shared/models folder
