+++ frontend/src/app/shared/models/credentials.model.ts
@@ -1,0 +1,4 @@
+export interface CredentialsModel {
+  login: string;
+  password: string;
+}
