+++ frontend/src/app/shared/models/error-dialog-data.model.ts
@@ -1,0 +1,3 @@
+export interface ErrorDialogDataModel {
+  statusCode: number;
+}
