+++ frontend/src/app/shared/models/error.model.ts
@@ -1,0 +1,6 @@
+export interface ErrorModel {
+  id: number;
+  message: string;
+  user_id: number;
+  service_id: number;
+}
