+++ frontend/src/app/shared/models/token-session.model.ts
@@ -1,0 +1,4 @@
+export class TokenSessionModel {
+  // tslint:disable-next-line:variable-name
+  token_session: string;
+}
